@@ -1,12 +1,20 @@
 [project]
-requires-python = "~=3.8"
+requires-python = "~=3.11"
 dependencies = [
-    "sphinx",
+    "sphinx>=8.1",
     "myst-parser",
     "sphinx-autobuild",
     "sphinx-inline-tabs",
     "sphinx-copybutton",
-    "furo",
+    "sphinxcontrib-autoprogram",
+    # If you change the version of Furo, check out the repository, apply the following patch
+    # to `src/furo/assets/styles/variables/_layout.scss`:
+    #   -$sidebar-width: 15em;
+    #   +$sidebar-width: 16em;
+    # then rebuild it and place the build product `src/furo/theme/furo/static/styles/furo.css`
+    # into `docs/manual/src/_static/styles/furo.css`.
+    "furo==2024.8.6",
+    "glasgow @ file:///${PROJECT_ROOT}/../../software"
 ]
 
 [tool.pdm.scripts]
