@@ -2,28 +2,29 @@ import logging
 import argparse
 from vcd import VCDWriter
 from amaranth import *
+from amaranth.lib import io
 from amaranth.lib.cdc import FFSynchronizer
 
-from ....gateware.pads import *
 from ....gateware.analyzer import *
 from ... import *
 
 
 class AnalyzerSubtarget(Elaboratable):
-    def __init__(self, pads, in_fifo):
-        self.pads    = pads
+    def __init__(self, ports, in_fifo):
+        self.ports   = ports
         self.in_fifo = in_fifo
 
         self.analyzer = EventAnalyzer(in_fifo)
-        self.event_source = self.analyzer.add_event_source("pin", "change", len(pads.i_t.i))
+        self.event_source = self.analyzer.add_event_source("pin", "change", len(self.ports.i))
 
     def elaborate(self, platform):
         m = Module()
         m.submodules += self.analyzer
 
-        pins_i = Signal.like(self.pads.i_t.i)
-        pins_r = Signal.like(self.pads.i_t.i)
-        m.submodules += FFSynchronizer(self.pads.i_t.i, pins_i)
+        m.submodules.i_buffer = i_buffer = io.Buffer("i", self.ports.i)
+        pins_i = Signal.like(i_buffer.i)
+        pins_r = Signal.like(i_buffer.i)
+        m.submodules += FFSynchronizer(i_buffer.i, pins_i)
 
         m.d.sync += pins_r.eq(pins_i)
         m.d.comb += [
@@ -56,12 +57,15 @@ class AnalyzerApplet(GlasgowApplet):
     def add_build_arguments(cls, parser, access):
         super().add_build_arguments(parser, access)
 
-        access.add_pin_set_argument(parser, "i", width=range(1, 17), default=1)
+        access.add_pins_argument(parser, "i", width=range(1, 17), default=1)
+        parser.add_argument(
+            "--pin-names", metavar="NAMES", dest="names", default=None,
+            help="optional comma separated list of pin names")
 
     def build(self, target, args):
         self.mux_interface = iface = target.multiplexer.claim_interface(self, args)
         subtarget = iface.add_subtarget(AnalyzerSubtarget(
-            pads=iface.get_pads(args, pin_sets=("i",)),
+            ports=iface.get_port_group(i = args.i),
             in_fifo=iface.get_in_fifo(),
         ))
 
@@ -84,9 +88,9 @@ class AnalyzerApplet(GlasgowApplet):
         pull_low  = set()
         pull_high = set()
         if args.pull_ups:
-            pull_high = set(args.pin_set_i)
+            pull_high = set(args.i)
         if args.pull_downs:
-            pull_low = set(args.pin_set_i)
+            pull_low = set(args.i)
         iface = await device.demultiplexer.claim_interface(self, self.mux_interface, args,
                                                            pull_low=pull_low, pull_high=pull_high)
         return AnalyzerInterface(iface, self._event_sources)
@@ -100,8 +104,16 @@ class AnalyzerApplet(GlasgowApplet):
     async def interact(self, device, args, iface):
         vcd_writer = VCDWriter(args.file, timescale="1 ns", check_values=False)
         signals = []
+
+        names = []
+        if args.names:
+            names = args.names.split(",")
+            assert len(names) == self._event_sources[0].width
+        else:
+            names = [ f"pin[{index}]" for index in range(self._event_sources[0].width) ]
+
         for index in range(self._event_sources[0].width):
-            signals.append(vcd_writer.register_var(scope="", name=f"pin[{index}]",
+            signals.append(vcd_writer.register_var(scope="glasgow", name=names[index],
                 var_type="wire", size=1, init=0))
 
         try:
