@@ -24,12 +24,11 @@ import logging
 import argparse
 import enum
 from amaranth import *
-from amaranth.lib.cdc import FFSynchronizer
+from amaranth.lib import io, cdc
 
 from ....support.bits import *
 from ....support.logging import *
 from ....support.arepl import *
-from ....gateware.pads import *
 from ....database.jedec import *
 from ....arch.jtag import *
 from ... import *
@@ -77,8 +76,9 @@ JTAG_TRANSITIONS = {
 
 
 class JTAGProbeBus(Elaboratable):
-    def __init__(self, pads):
-        self._pads = pads
+    def __init__(self, ports):
+        self._ports = ports
+
         self.tck = Signal(init=1)
         self.tms = Signal(init=1)
         self.tdo = Signal(init=1)
@@ -88,23 +88,18 @@ class JTAGProbeBus(Elaboratable):
 
     def elaborate(self, platform):
         m = Module()
-        pads = self._pads
-        m.d.comb += [
-            pads.tck_t.oe.eq(1),
-            pads.tck_t.o.eq(self.tck),
-            pads.tms_t.oe.eq(1),
-            pads.tms_t.o.eq(self.tms),
-            pads.tdi_t.oe.eq(1),
-            pads.tdi_t.o.eq(self.tdi),
-        ]
-        m.submodules += [
-            FFSynchronizer(pads.tdo_t.i, self.tdo),
-        ]
-        if hasattr(pads, "trst_t"):
-            m.d.sync += [
-                pads.trst_t.oe.eq(~self.trst_z),
-                pads.trst_t.o.eq(~self.trst_o)
-            ]
+        m.submodules.tck = tck_buffer = io.Buffer("o", self._ports.tck)
+        m.d.comb += tck_buffer.o.eq(self.tck)
+        m.submodules.tms = tms_buffer = io.Buffer("o", self._ports.tms)
+        m.d.comb += tms_buffer.o.eq(self.tms)
+        m.submodules.tdi = tdi_buffer = io.Buffer("o", self._ports.tdi)
+        m.d.comb += tdi_buffer.o.eq(self.tdi)
+        m.submodules.tdo = tdo_buffer = io.Buffer("i", self._ports.tdo)
+        m.submodules += cdc.FFSynchronizer(tdo_buffer.i, self.tdo)
+        if self._ports.trst is not None:
+            m.submodules.trst = trst_buffer = io.Buffer("o", ~self._ports.trst)
+            m.d.comb += trst_buffer.oe.eq(~self.trst_z)
+            m.d.comb += trst_buffer.o.eq(self.trst_o)
         return m
 
 
@@ -292,15 +287,15 @@ class JTAGProbeDriver(Elaboratable):
 
 
 class JTAGProbeSubtarget(Elaboratable):
-    def __init__(self, pads, out_fifo, in_fifo, period_cyc):
-        self._pads       = pads
+    def __init__(self, ports, out_fifo, in_fifo, period_cyc):
+        self._ports      = ports
         self._out_fifo   = out_fifo
         self._in_fifo    = in_fifo
         self._period_cyc = period_cyc
 
     def elaborate(self, platform):
         m = Module()
-        m.submodules.bus     = JTAGProbeBus(self._pads)
+        m.submodules.bus     = JTAGProbeBus(self._ports)
         m.submodules.adapter = JTAGProbeAdapter(m.submodules.bus, self._period_cyc)
         m.submodules.driver  = JTAGProbeDriver(m.submodules.adapter, self._out_fifo, self._in_fifo)
         return m
@@ -1005,15 +1000,13 @@ class JTAGProbeApplet(GlasgowApplet):
     Identify, test and debug integrated circuits and board assemblies via IEEE 1149.1 JTAG.
     """
 
-    __pins = ("tck", "tms", "tdi", "tdo", "trst")
-
     @classmethod
     def add_build_arguments(cls, parser, access):
         super().add_build_arguments(parser, access)
 
         for pin in ("tck", "tms", "tdi", "tdo"):
-            access.add_pin_argument(parser, pin, default=True)
-        access.add_pin_argument(parser, "trst")
+            access.add_pins_argument(parser, pin, default=True)
+        access.add_pins_argument(parser, "trst")
 
         parser.add_argument(
             "-f", "--frequency", metavar="FREQ", type=int, default=100,
@@ -1022,7 +1015,13 @@ class JTAGProbeApplet(GlasgowApplet):
     def build(self, target, args):
         self.mux_interface = iface = target.multiplexer.claim_interface(self, args)
         iface.add_subtarget(JTAGProbeSubtarget(
-            pads=iface.get_pads(args, pins=self.__pins),
+            ports=iface.get_port_group(
+                tck=args.tck,
+                tms=args.tms,
+                tdi=args.tdi,
+                tdo=args.tdo,
+                trst=args.trst,
+            ),
             out_fifo=iface.get_out_fifo(),
             in_fifo=iface.get_in_fifo(auto_flush=False),
             period_cyc=target.sys_clk_freq // (args.frequency * 1000),
@@ -1061,7 +1060,7 @@ class JTAGProbeApplet(GlasgowApplet):
 
     async def run(self, device, args):
         iface = await device.demultiplexer.claim_interface(self, self.mux_interface, args)
-        jtag_iface = JTAGProbeInterface(iface, self.logger, has_trst=args.pin_trst is not None)
+        jtag_iface = JTAGProbeInterface(iface, self.logger, has_trst=args.trst is not None)
         jtag_iface.scan_ir_max_length = args.scan_ir_max_length
         jtag_iface.scan_dr_max_length = args.scan_dr_max_length
         return jtag_iface
