@@ -1,5 +1,6 @@
 import types
 from amaranth import *
+from amaranth.lib import io
 
 from ... import *
 from . import SPIControllerApplet
@@ -8,26 +9,26 @@ from . import SPIControllerApplet
 class SPIControllerAppletTestCase(GlasgowAppletTestCase, applet=SPIControllerApplet):
     @synthesis_test
     def test_build(self):
-        self.assertBuilds(args=["--pin-sck",  "0", "--pin-cs",   "1",
-                                "--pin-copi", "2", "--pin-cipo", "3"])
+        self.assertBuilds(args=["--sck",  "A0", "--cs",   "A1",
+                                "--copi", "A2", "--cipo", "A3"])
 
-    def setup_loopback(self):
-        self.build_simulated_applet()
-        mux_iface = self.applet.mux_interface
-        m = Module()
-        m.d.comb += mux_iface.pads.cipo_t.i.eq(mux_iface.pads.copi_t.o)
-        self.target.add_submodule(m)
+    def setup_loopback(self, target, parsed_args):
+        self.applet.build(target, parsed_args)
+        target.assembly.connect_pins("A2", "A3")
 
     @applet_simulation_test("setup_loopback",
-                            ["--pin-sck",  "0", "--pin-cs", "1",
-                             "--pin-copi", "2", "--pin-cipo",   "3",
-                             "--frequency", "5000"])
-    @types.coroutine
-    def test_loopback(self):
-        mux_iface = self.applet.mux_interface
-        spi_iface = yield from self.run_simulated_applet()
+                            ["--sck",  "A0", "--cs",   "A1",
+                             "--copi", "A2", "--cipo", "A3",
+                             "--frequency", "10"])
+    async def test_loopback(self, device, parsed_args, ctx):
+        spi_iface = await self.applet.run(device, parsed_args)
 
-        self.assertEqual((yield mux_iface.pads.cs_t.o), 1)
-        result = yield from spi_iface.transfer([0xAA, 0x55, 0x12, 0x34])
-        self.assertEqual(result, bytearray([0xAA, 0x55, 0x12, 0x34]))
-        self.assertEqual((yield mux_iface.pads.cs_t.o), 1)
+        cs = device.assembly.get_pin("A1")
+        self.assertEqual(ctx.get(cs.o), 1)
+        async with spi_iface.select():
+            await ctx.tick().repeat(2)
+            self.assertEqual(ctx.get(cs.o), 0)
+            result = await spi_iface.exchange([0xAA, 0x55, 0x12, 0x34])
+            self.assertEqual(result, bytearray([0xAA, 0x55, 0x12, 0x34]))
+        await ctx.tick()
+        self.assertEqual(ctx.get(cs.o), 1)
