@@ -4,20 +4,21 @@ from ... import *
 from . import UARTApplet
 
 
-class UARTAppletTestCase(GlasgowAppletTestCase, applet=UARTApplet):
+class UARTAppletTestCase(GlasgowAppletV2TestCase, applet=UARTApplet):
     @synthesis_test
     def test_build(self):
         self.assertBuilds()
 
-    def setup_loopback(self):
-        self.build_simulated_applet()
-        mux_iface = self.applet.mux_interface
-        m = Module()
-        m.d.comb += mux_iface.pads.rx_t.i.eq(mux_iface.pads.tx_t.o)
-        self.target.add_submodule(m)
+    def prepare_loopback(self, assembly):
+        assembly.connect_pins("A0", "A1")
 
-    @applet_simulation_test("setup_loopback", ["--baud", "5000000"])
-    async def test_loopback(self):
-        uart_iface = await self.run_simulated_applet()
-        await uart_iface.write(bytes([0xAA, 0x55]))
-        self.assertEqual(await uart_iface.read(2), bytes([0xAA, 0x55]))
+    @applet_v2_simulation_test(prepare=prepare_loopback, args="--baud 9600")
+    async def test_loopback(self, applet, ctx):
+        await applet.uart_iface.write(bytes([0xAA, 0x55]))
+        self.assertEqual(await applet.uart_iface.read(2), bytes([0xAA, 0x55]))
+
+    # This test is here mainly to test the test machinery.
+    @applet_v2_hardware_test(args="-V 3.3 --baud 9600", mock="uart_iface")
+    async def test_loopback_hw(self, applet):
+        await applet.uart_iface.write(bytes([0xAA, 0x55]))
+        self.assertEqual(await applet.uart_iface.read(2), bytes([0xAA, 0x55]))
