@@ -2,7 +2,7 @@ from ... import *
 from . import ProgramICE40FlashApplet
 
 
-class ProgramICE40FlashAppletTestCase(GlasgowAppletTestCase, applet=ProgramICE40FlashApplet):
+class ProgramICE40FlashAppletTestCase(GlasgowAppletV2TestCase, applet=ProgramICE40FlashApplet):
     @synthesis_test
     def test_build(self):
-        self.assertBuilds()
+        self.assertBuilds("--reset B0 --done B1")
