@@ -1,32 +1,40 @@
+# Ref: iCE40 Programming and Configuration Technical Note
+# Document Number: FPGA-TN-02001-3.2
+# Accession: G00073
+
 import argparse
 import asyncio
 import logging
 from amaranth import *
+from amaranth.lib import io
 
 from ...interface.spi_controller import SPIControllerApplet
 from ... import *
 
 
 class ProgramICE40SRAMSubtarget(Elaboratable):
-    def __init__(self, controller, reset_t, dut_reset, done_t, dut_done):
+    def __init__(self, controller, port_reset, dut_reset, port_done, dut_done):
         self.controller = controller
-        self.reset_t = reset_t
-        self.dut_reset = dut_reset
-        self.done_t = done_t
-        self.dut_done = dut_done
+        self.port_reset = port_reset
+        self.dut_reset  = dut_reset
+        self.port_done  = port_done
+        self.dut_done   = dut_done
 
     def elaborate(self, platform):
         m = Module()
 
         m.submodules.controller = self.controller
 
+        m.submodules.reset_buffer = reset_buffer = io.Buffer("o", self.port_reset)
+
         m.d.comb += [
-            self.reset_t.o.eq(0),
-            self.reset_t.oe.eq(self.dut_reset)
+            reset_buffer.o.eq(0),
+            reset_buffer.oe.eq(self.dut_reset)
         ]
 
-        if self.done_t is not None:
-            m.d.comb += self.dut_done.eq(self.done_t.i)
+        if self.port_done is not None:
+            m.submodules.done_buffer = done_buffer = io.Buffer("i", self.port_done)
+            m.d.comb += self.dut_done.eq(done_buffer.i)
 
         return m
 
@@ -44,27 +52,27 @@ class ProgramICE40SRAMInterface:
         await self._device.write_register(self._addr_dut_reset, int(reset))
 
     async def get_done(self):
-        if self._addr_dut_reset is not None:
+        if self._addr_dut_done is not None:
             return await self._device.read_register(self._addr_dut_done)
+        else:
+            return True
 
     async def program(self, bitstream):
-        # Pulse reset while setting SS_n low (we use a dummy write to do that)
-        await self.set_reset(True)
-        await self.lower.write([], hold_ss=True)
-        await self.lower.synchronize()
-        await self.set_reset(False)
+        async with self.lower.select():
+            # Pulse reset while holding SSn low
+            await self.lower.synchronize()
+            await self.set_reset(True)
+            await self.set_reset(False)
+            await self.lower.synchronize()
 
-        # Wait at least 1.2ms (spec for 8k devices)
-        await self.lower.delay_us(1200)
+            # Wait at least 1.2ms (spec for 8k devices)
+            await self.lower.delay_us(1200)
 
-        # Write bitstream
-        while len(bitstream) > 0:
-            chunk = bitstream[:255]
-            bitstream = bitstream[255:]
-            await self.lower.write(chunk, hold_ss=True)
+            # Write bitstream
+            await self.lower.write(bitstream)
 
-        # Specs says at least 49 dummy bits. Send 128.
-        await self.lower.write([0] * 16)
+            # Specs says at least 49 dummy bits. Send 128.
+            await self.lower.dummy(128)
 
 
 class ProgramICE40SRAMApplet(SPIControllerApplet):
@@ -78,27 +86,27 @@ class ProgramICE40SRAMApplet(SPIControllerApplet):
     def add_build_arguments(cls, parser, access):
         super().add_build_arguments(parser, access, omit_pins=True)
 
-        access.add_pin_argument(parser, "sck",   required=True)
-        access.add_pin_argument(parser, "cs",    required=True)
-        access.add_pin_argument(parser, "copi",  required=True)
-        access.add_pin_argument(parser, "reset", required=True)
-        access.add_pin_argument(parser, "done")
+        access.add_pins_argument(parser, "sck",   required=True)
+        access.add_pins_argument(parser, "cs",    required=True)
+        access.add_pins_argument(parser, "copi",  required=True)
+        access.add_pins_argument(parser, "reset", required=True)
+        access.add_pins_argument(parser, "done")
 
     def build_subtarget(self, target, args):
-        subtarget = super().build_subtarget(target, args, pins=("sck", "cs", "copi"))
+        subtarget = super().build_subtarget(target, args)
 
-        reset_t = self.mux_interface.get_pin(args.pin_reset)
+        port_reset = self.mux_interface.get_port(args.reset, name="reset")
         dut_reset, self.__addr_dut_reset = target.registers.add_rw(1)
 
-        if args.pin_done is not None:
-            done_t = self.mux_interface.get_pin(args.pin_done)
+        if args.done is not None:
+            port_done = self.mux_interface.get_port(args.done, name="done")
             dut_done, self.__addr_dut_done = target.registers.add_ro(1)
         else:
-            done_t = None
+            port_done = None
             dut_done = None
             self.__addr_dut_done = None
 
-        return ProgramICE40SRAMSubtarget(subtarget, reset_t, dut_reset, done_t, dut_done)
+        return ProgramICE40SRAMSubtarget(subtarget, port_reset, dut_reset, port_done, dut_done)
 
     @classmethod
     def add_interact_arguments(cls, parser):
@@ -115,7 +123,7 @@ class ProgramICE40SRAMApplet(SPIControllerApplet):
         bitstream = args.bitstream.read()
         await ice40_iface.program(bitstream)
 
-        if args.pin_done is not None:
+        if args.done is not None:
             for _ in range(200):    # Wait up to 2s
                 await asyncio.sleep(0.010)  # Poll every 10 ms
                 done = await ice40_iface.get_done()
