@@ -5,4 +5,4 @@ from . import VGAOutputApplet
 class VGAOutputAppletTestCase(GlasgowAppletTestCase, applet=VGAOutputApplet):
     @synthesis_test
     def test_build(self):
-        self.assertBuilds(args=["--port", "B"])
+        self.assertBuilds()
