@@ -1,38 +1,35 @@
 import logging
 import asyncio
 from amaranth import *
+from amaranth.lib import io
 
 from ....support.endpoint import *
-from ....gateware.pads import *
 from ....gateware.pll import *
 from ... import *
 
 
 class VideoWS2812Output(Elaboratable):
-    def __init__(self, pads):
-        self.pads = pads
-        self.out = Signal(len(pads))
+    def __init__(self, ports):
+        self.ports = ports
+        self.out = Signal(len(ports.out))
 
     def elaborate(self, platform):
         m = Module()
 
-        for i, pad in enumerate(self.pads):
-            m.d.comb += [
-                pad.oe.eq(1),
-                pad.o.eq(self.out[i]),
-            ]
+        m.submodules.out_buffer = out_buffer = io.Buffer("o", self.ports.out)
+        m.d.comb += out_buffer.o.eq(self.out)
 
         return m
 
 
 class VideoWS2812OutputSubtarget(Elaboratable):
-    def __init__(self, pads, count, pix_in_size, pix_out_size, pix_format_func, out_fifo):
-        self.pads = pads
-        self.count = count
-        self.pix_in_size = pix_in_size
-        self.pix_out_size = pix_out_size
+    def __init__(self, ports, count, pix_in_size, pix_out_size, pix_format_func, out_fifo):
+        self.ports           = ports
+        self.count           = count
+        self.pix_in_size     = pix_in_size
+        self.pix_out_size    = pix_out_size
         self.pix_format_func = pix_format_func
-        self.out_fifo = out_fifo
+        self.out_fifo        = out_fifo
 
     def elaborate(self, platform):
         # Safe timings:
@@ -51,7 +48,7 @@ class VideoWS2812OutputSubtarget(Elaboratable):
 
         m = Module()
 
-        m.submodules.output = output = VideoWS2812Output(self.pads)
+        m.submodules.output = output = VideoWS2812Output(self.ports)
 
         pix_in_size = self.pix_in_size
         pix_out_size = self.pix_out_size
@@ -61,10 +58,10 @@ class VideoWS2812OutputSubtarget(Elaboratable):
         bit_ctr = Signal(range(pix_out_bpp+1))
         byt_ctr = Signal(range((pix_in_size)+1))
         pix_ctr = Signal(range(self.count+1))
-        word_ctr = Signal(range(max(2, len(self.pads))))
+        word_ctr = Signal(range(max(2, len(self.ports.out))))
 
         pix = Array([ Signal(8) for i in range((pix_in_size) - 1) ])
-        word = Signal(pix_out_bpp * len(self.pads))
+        word = Signal(pix_out_bpp * len(self.ports.out))
 
         with m.FSM():
             with m.State("LOAD"):
@@ -81,7 +78,7 @@ class VideoWS2812OutputSubtarget(Elaboratable):
                     with m.Else():
                         p = self.pix_format_func(*pix, self.out_fifo.r_data)
                         m.d.sync += word.eq(Cat(word[pix_out_bpp:], p))
-                        with m.If(word_ctr < (len(self.pads) - 1)):
+                        with m.If(word_ctr < (len(self.ports.out) - 1)):
                             m.d.sync += [
                                 word_ctr.eq(word_ctr + 1),
                                 byt_ctr.eq(0),
@@ -91,7 +88,7 @@ class VideoWS2812OutputSubtarget(Elaboratable):
 
             with m.State("SEND-WORD"):
                 with m.If(cyc_ctr < t_zero):
-                    m.d.comb += output.out.eq((1 << len(self.pads)) - 1)
+                    m.d.comb += output.out.eq((1 << len(self.ports.out)) - 1)
                     m.d.sync += cyc_ctr.eq(cyc_ctr + 1)
                 with m.Elif(cyc_ctr < t_one):
                     m.d.comb += ( o.eq(word[(pix_out_bpp - 1) + (pix_out_bpp * i)]) for i,o in enumerate(output.out) )
@@ -155,7 +152,7 @@ class VideoWS2812OutputApplet(GlasgowApplet):
     def add_build_arguments(cls, parser, access):
         super().add_build_arguments(parser, access)
 
-        access.add_pin_set_argument(parser, "out", width=range(1, 17), required=True)
+        access.add_pins_argument(parser, "out", width=range(1, 17), required=True)
         parser.add_argument(
             "-c", "--count", metavar="N", type=int, required=True,
             help="set the number of LEDs per string")
@@ -168,7 +165,7 @@ class VideoWS2812OutputApplet(GlasgowApplet):
 
         self.mux_interface = iface = target.multiplexer.claim_interface(self, args)
         subtarget = iface.add_subtarget(VideoWS2812OutputSubtarget(
-            pads=[iface.get_pin(pin) for pin in args.pin_set_out],
+            ports=iface.get_port_group(out=args.out),
             count=args.count,
             pix_in_size=self.pix_in_size,
             pix_out_size=pix_out_size,
@@ -187,7 +184,7 @@ class VideoWS2812OutputApplet(GlasgowApplet):
             help="set the number of frames to buffer internally (buffered twice)")
 
     async def run(self, device, args):
-        buffer_size = len(args.pin_set_out) * args.count * self.pix_in_size * args.buffer
+        buffer_size = len(args.out) * args.count * self.pix_in_size * args.buffer
         return await device.demultiplexer.claim_interface(self, self.mux_interface, args, write_buffer_size=buffer_size)
 
     @classmethod
@@ -195,9 +192,10 @@ class VideoWS2812OutputApplet(GlasgowApplet):
         ServerEndpoint.add_argument(parser, "endpoint")
 
     async def interact(self, device, args, leds):
-        frame_size = len(args.pin_set_out) * args.count * self.pix_in_size
+        frame_size = len(args.out) * args.count * self.pix_in_size
         buffer_size = frame_size * args.buffer
-        endpoint = await ServerEndpoint("socket", self.logger, args.endpoint, queue_size=buffer_size)
+        endpoint = await ServerEndpoint("socket", self.logger, args.endpoint, queue_size=buffer_size,
+            deprecated_cancel_on_eof=True)
         while True:
             try:
                 data = await asyncio.shield(endpoint.recv(buffer_size))
