@@ -143,213 +143,117 @@
 # FIFOADR->FIFODATA             14.3
 
 from amaranth import *
-from amaranth.lib.cdc import FFSynchronizer, ResetSynchronizer
-from amaranth.lib.fifo import FIFOInterface, AsyncFIFO, SyncFIFO, SyncFIFOBuffered
-from amaranth.lib import io
+from amaranth.lib import wiring, stream, io
+from amaranth.lib.wiring import In, Out, connect, flipped
+
+from .stream import StreamFIFO
 
 
 __all__ = ["FX2Crossbar"]
 
 
-class _OUTFIFO(Elaboratable, FIFOInterface):
-    """
-    A FIFO with a skid buffer in front of it. This FIFO may be fed from a pipeline that
-    reacts to the ``w_rdy`` flag with a latency up to the skid buffer depth, and writes
-    will not be lost.
-    """
-    def __init__(self, inner, skid_depth):
-        super().__init__(width=inner.width, depth=inner.depth)
-
-        self.inner = inner
-        self.skid  = skid  = SyncFIFO(width=inner.width, depth=skid_depth)
-
-        self.r_data   = inner.r_data
-        self.r_en     = inner.r_en
-        self.r_rdy    = inner.r_rdy
-
-    def elaborate(self, platform):
-        m = Module()
-
-        m.submodules.inner = inner = self.inner
-        m.submodules.skid = skid = self.skid
-
-        with m.If(skid.r_rdy):
-            m.d.comb += [
-                inner.w_data.eq(skid.r_data),
-                inner.w_en.eq(1),
-                skid.r_en.eq(inner.w_rdy)
-            ]
-
-        with m.If(inner.w_rdy & ~skid.r_rdy):
-            m.d.comb += [
-                inner.w_data.eq(self.w_data),
-                inner.w_en.eq(self.w_en),
-                self.w_rdy.eq(inner.w_rdy)
-            ]
-        with m.Else():
-            m.d.comb += [
-                skid.w_data.eq(self.w_data),
-                skid.w_en.eq(self.w_en),
-                self.w_rdy.eq(skid.w_rdy)
-            ]
-
-        return m
+_PACKET_SIZE = 512
 
 
-class _UnimplementedOUTFIFO(FIFOInterface):
-    def __init__(self, width):
-        super().__init__(width=width, depth=0)
-
-        self.inner = FIFOInterface(width=width, depth=0)
-
-
-class _INFIFO(Elaboratable, FIFOInterface):
+class _INFIFO(wiring.Component):
     """
     A FIFO with a sideband flag indicating whether the FIFO has enough data to read from it yet.
     This FIFO may be used for packetizing the data read from the FIFO when there is no particular
     framing available to optimize the packet boundaries.
     """
-    def __init__(self, inner, packet_size=512, asynchronous=False, auto_flush=True):
-        super().__init__(width=inner.width, depth=inner.depth)
 
-        self.inner = inner
-        self.packet_size = packet_size
-        self.asynchronous = asynchronous
-        self.auto_flush = auto_flush
+    w: In(stream.Signature(8))
+    r: Out(stream.Signature(8))
+    flush: In(1)
 
-        self.r_data   = inner.r_data
-        self.r_en     = inner.r_en
-        self.r_rdy    = inner.r_rdy
-        self.w_data   = inner.w_data
-        self.w_en     = inner.w_en
-        self.w_rdy    = inner.w_rdy
-
-        self.flush = Signal(init=auto_flush)
-
-        # This is a model of the IN FIFO buffer in the FX2. Keep in mind that it is legal
-        # to assert PKTEND together with SLWR, and in that case PKTEND takes priority.
-        # This model is placed in the _INFIFO so that it is reset together with the FIFO itself,
-        # which happens on Set Configuration and Set Interface requests.
-        self.queued   = Signal(range(1 + packet_size))
-        self.complete = Signal() # one FX2 FIFO buffer full
-        self.pending  = Signal() # PKTEND requested
-        self.flushed  = Signal() # PKTEND asserted
+    # This is a model of the IN FIFO buffer in the FX2. Keep in mind that it is legal to assert
+    # PKTEND together with SLWR, and in that case PKTEND takes priority. Also, note that this
+    # model must be reset on Set Configuration and Set Interface requests.
+    queued:   Out(range(_PACKET_SIZE + 1))
+    complete: Out(1) # one FX2 FIFO buffer full
+    pending:  Out(1) # PKTEND requested
+    flushed:  In(1)  # PKTEND asserted
 
     def elaborate(self, platform):
         m = Module()
 
-        m.submodules.inner = self.inner
+        connect(m, flipped(self.r), flipped(self.w))
 
-        if self.asynchronous:
-            _flush_s = Signal()
-            m.submodules += FFSynchronizer(self.flush, _flush_s, init=self.auto_flush)
-        else:
-            _flush_s = self.flush
-
-        _pending = Signal()
+        pending = Signal()
         with m.If(self.flushed):
             m.d.sync += self.queued.eq(0)
             # If we sent a maximum-size packet, we still need a ZLP afterwards.
-            with m.If(self.queued < self.packet_size):
-                m.d.sync += _pending.eq(0)
-        with m.Elif(self.r_rdy & self.r_en):
+            with m.If(self.queued < _PACKET_SIZE):
+                m.d.sync += pending.eq(0)
+        with m.Elif(self.r.valid & self.r.ready):
             m.d.sync += [
                 self.queued.eq(self.queued + 1),
-                _pending.eq(1)
+                pending.eq(1)
             ]
 
         m.d.comb += [
-            self.complete.eq(self.queued >= self.packet_size),
-            self.pending.eq(_pending & _flush_s),
+            self.complete.eq(self.queued >= _PACKET_SIZE),
+            self.pending.eq(pending & self.flush),
         ]
 
         return m
 
 
-class _UnimplementedINFIFO(FIFOInterface):
-    def __init__(self, width, packet_size=512):
-        super().__init__(width=width, depth=0)
+class _OUTFIFO(wiring.Component):
+    """
+    A FIFO with a skid buffer in front of it. This FIFO may be fed from a pipeline that
+    reacts to the ``w.ready`` flag with a latency up to the skid buffer depth, and writes
+    will not be lost.
 
-        self.inner = FIFOInterface(width=width, depth=0)
+    Note that the ``w`` interface does not conform to the usual stream invariants as a result.
+    """
 
-        self.flush    = Signal()
-        self.flush_s  = Signal()
+    w: In(stream.Signature(8))
+    r: Out(stream.Signature(8))
 
-        self.queued   = Signal(range(1 + packet_size))
-        self.complete = Signal()
-        self.pending  = Signal()
-        self.flushed  = Signal()
+    def __init__(self, skid_depth):
+        self._skid_depth = skid_depth
 
-
-class _AsyncFIFOWrapper(Elaboratable, FIFOInterface):
-    def __init__(self, inner, cd_logic, reset):
-        super().__init__(width=inner.width, depth=inner.depth)
-
-        self.inner = inner
-        self.cd_logic = cd_logic
-        self.reset = reset
-
-        self.r_data   = inner.r_data
-        self.r_en     = inner.r_en
-        self.r_rdy    = inner.r_rdy
-        self.w_data   = inner.w_data
-        self.w_en     = inner.w_en
-        self.w_rdy    = inner.w_rdy
+        super().__init__()
 
     def elaborate(self, platform):
         m = Module()
 
-        m.submodules.inner = self.inner
+        m.submodules.skid = skid = StreamFIFO(shape=8, depth=self._skid_depth, buffered=False)
 
-        # Note that for the reset to get asserted AND deasserted, the logic clock domain must
-        # have a running clock. This is because, while AsyncResetSynchronizer is indeed
-        # asynchronous, the registers in the FIFO logic clock domain reset synchronous
-        # to the logic clock, as this is how Migen handles clock domain reset signals.
-        #
-        # If the logic clock domain does not have a single clock transition between assertion
-        # and deassertion of FIFO reset, and the FIFO has not been empty at the time when
-        # reset has been asserted, stale data will be read from the FIFO after deassertion.
-        #
-        # This can lead to all sorts of framing issues, and is rather unfortunate, but at
-        # the moment I do not know of a way to fix this, since Migen does not support
-        # asynchronous resets.
-        cd_crossbar = ClockDomain(reset_less=self.reset is None, local=True)
-        cd_logic    = ClockDomain(reset_less=self.reset is None, local=True)
-        m.d.comb += [
-            cd_crossbar.clk.eq(ClockSignal()),
-            cd_logic.clk.eq(self.cd_logic.clk),
-        ]
-        if self.reset is not None:
-            m.d.comb += cd_crossbar.rst.eq(self.reset)
-            m.submodules += ResetSynchronizer(self.reset, domain="logic")
-
-        m.domains.logic = cd_logic
-        m.domains.crossbar = cd_crossbar
+        m.d.comb += skid.w.payload.eq(self.w.payload)
+        m.d.comb += skid.w.valid.eq(self.w.valid & (~self.r.ready | skid.r.valid))
+        with m.If(skid.r.valid):
+            connect(m, flipped(self.r), skid.r)
+        with m.Else():
+            connect(m, flipped(self.r), flipped(self.w))
 
         return m
 
 
-class _FX2Bus(Elaboratable):
+class _FX2Bus(wiring.Component):
+    flag: Out(4)
+    addr: In(2)
+    data: In(io.Buffer.Signature("io", 8))
+    sloe: In(1)
+    slrd: In(1)
+    slwr: In(1)
+    pend: In(1)
+
+    addr_p: Out(2)
+    slrd_p: Out(1)
+
+    # When an operation is pipelined that may or may not change flags, it is useful to
+    # invalidate--artificially negate--the corresponding flag until the operation completes.
+    # The nrdy signals are delayed by _FX2Bus in the same way as other pipelined signals,
+    # and an output bit is active while any corresponding bit in the pipeline is still active.
+    nrdy_i: In(4)
+    nrdy_o: Out(4)
+
     def __init__(self, pads):
-        self.flag = Signal(4)
-        self.addr = Signal(2)
-        self.data = io.Buffer.Signature('io', 8).create()
-        self.sloe = Signal()
-        self.slrd = Signal()
-        self.slwr = Signal()
-        self.pend = Signal()
-
-        self.addr_p = Signal.like(self.addr)
-        self.slrd_p = Signal.like(self.slrd)
-
-        # When an operation is pipelined that may or may not change flags, it is useful to
-        # invalidate--artificially negate--the corresponding flag until the operation completes.
-        # The nrdy signals are delayed by _FX2Bus in the same way as other pipelined signals,
-        # and an output bit is active while any corresponding bit in the pipeline is still active.
-        self.nrdy_i = Signal.like(self.flag)
-        self.nrdy_o = Signal.like(self.flag)
-
         self.pads = pads
+
+        super().__init__()
 
     def elaborate(self, platform):
         m = Module()
@@ -400,62 +304,71 @@ class _FX2Bus(Elaboratable):
         return m
 
 
-class FX2Crossbar(Elaboratable):
-    """
-    FX2 FIFO bus master.
+class FX2Crossbar(wiring.Component):
+    """FX2 bus to FIFO crossbar.
 
-    Shuttles data between FX2 and FIFOs in bursts.
-
-    The crossbar supports up to four FIFOs organized as ``OUT, OUT, IN, IN``.
-    FIFOs that are never requested are not implemented and behave as if they
-    are never readable or writable.
+    The crossbar addresses four FX2 endpoints: two ``OUT`` followed by two ``IN``, in this order.
+    On the FPGA side, it provides one stream per endpoint, instantiating only a minimal amount of
+    logic necessary to coordinate the transfers. The FIFOs must be instantiated externally to
+    the crossbar.
     """
+
+    in_eps: Out(wiring.Signature({
+        "data":  In(stream.Signature(8)),
+        "flush": In(1),
+        "reset": In(1, reset=1),
+    })).array(2)
+    out_eps: Out(wiring.Signature({
+        "data":  Out(stream.Signature(8)),
+        "reset": In(1, reset=1),
+    })).array(2)
+
     def __init__(self, pads):
-        self.bus = _FX2Bus(pads)
+        self._pads = pads
 
-        self.out_fifos = Array([_UnimplementedOUTFIFO(width=8)
-                                for _ in range(2)])
-        self. in_fifos = Array([_UnimplementedINFIFO(width=8)
-                                for _ in range(2)])
-
+        super().__init__()
 
     def elaborate(self, platform):
         m = Module()
 
-        m.submodules.bus = bus = self.bus
+        m.submodules.bus = bus = _FX2Bus(self._pads)
 
-        for n, fifo in enumerate(self.in_fifos):
-            if isinstance(fifo, Elaboratable):
-                m.submodules[f"in_fifo_{n}"] = fifo
-        for n, fifo in enumerate(self.out_fifos):
-            if isinstance(fifo, Elaboratable):
-                m.submodules[f"out_fifo_{n}"] = fifo
+        in_fifos = Array([_INFIFO() for _ in self. in_eps])
+        for idx, (in_fifo, in_ep) in enumerate(zip(in_fifos, self.in_eps)):
+            m.submodules[ f"in_fifo_{idx}"] = ResetInserter(in_ep.reset)(in_fifo)
+            connect(m, in_fifo.w, flipped(in_ep.data))
+            m.d.comb += in_fifo.flush.eq(in_ep.flush)
+
+        out_fifos = Array([_OUTFIFO(skid_depth=3) for _ in self.out_eps])
+        for idx, (out_fifo, out_ep) in enumerate(zip(out_fifos, self.out_eps)):
+            m.submodules[f"out_fifo_{idx}"] = ResetInserter(out_ep.reset)(out_fifo)
+            connect(m, out_fifo.r, flipped(out_ep.data))
 
         rdy = Signal(4)
         m.d.comb += [
-            rdy.eq(Cat([fifo.inner.w_rdy          for fifo in self.out_fifos] +
-                       [fifo.r_rdy | fifo.pending for fifo in self. in_fifos]) &
+            rdy.eq(Cat([fifo.w.ready                for fifo in out_fifos] +
+                       [fifo.r.valid | fifo.pending for fifo in  in_fifos]) &
                    bus.flag &
                    ~bus.nrdy_o),
         ]
 
         sel_flag     = bus.flag.bit_select(bus.addr, 1)
-        sel_in_fifo  = self.in_fifos [bus.addr  [0]]
-        sel_out_fifo = self.out_fifos[bus.addr_p[0]]
+        sel_in_fifo  =  in_fifos[bus.addr  [0]]
+        sel_out_fifo = out_fifos[bus.addr_p[0]]
         m.d.comb += [
-            bus.data.o.eq(sel_in_fifo.r_data),
-            sel_out_fifo.w_data.eq(bus.data.i),
+            bus.data.o.eq(sel_in_fifo.r.payload),
+            sel_out_fifo.w.payload.eq(bus.data.i),
         ]
 
         with m.If(bus.addr[1]):
             m.d.comb += [
-                sel_in_fifo.r_en.eq(bus.slwr),
+                sel_in_fifo.r.ready.eq(bus.slwr),
                 sel_in_fifo.flushed.eq(bus.pend),
                 bus.nrdy_i.eq(Cat(C(0b00, 2), (bus.slwr | bus.pend) << bus.addr[0])),
             ]
         with m.Else():
             m.d.comb += [
-                sel_out_fifo.w_en.eq(bus.slrd_p & sel_flag),
+                sel_out_fifo.w.valid.eq(bus.slrd_p & sel_flag),
             ]
 
         # The FX2 requires the following setup latencies in worst case:
@@ -482,76 +395,33 @@ class FX2Crossbar(Elaboratable):
                                 m.d.sync += bus.addr.eq(addr_n)
                 with m.If(rdy):
                     m.next = "DRIVE"
+
             with m.State("DRIVE"):
                 with m.If(bus.addr[1]):
                     m.d.sync += bus.data.oe.eq(1)
                 with m.Else():
                     m.d.sync += bus.sloe.eq(1)
                 m.next = "SETUP"
+
             with m.State("SETUP"):
                 with m.If(bus.addr[1]):
                     m.next = "IN-XFER"
                 with m.Else():
                     m.next = "OUT-XFER"
+
             with m.State("IN-XFER"):
-                with m.If(~sel_in_fifo.complete & sel_in_fifo.r_rdy):
+                with m.If(~sel_in_fifo.complete & sel_in_fifo.r.valid):
                     m.d.comb += bus.slwr.eq(1)
                 with m.Elif(sel_in_fifo.complete | sel_in_fifo.pending):
                     m.d.comb += bus.pend.eq(1)
                     m.next = "SWITCH"
                 with m.Else():
                     m.next = "SWITCH"
+
             with m.State("OUT-XFER"):
-                with m.If(sel_flag & sel_out_fifo.inner.w_rdy):
+                with m.If(sel_flag & sel_out_fifo.w.ready):
                     m.d.comb += bus.slrd.eq(1)
                 with m.Else():
                     m.next = "SWITCH"
 
         return m
-
-    def _make_fifo(self, crossbar_side, logic_side, cd_logic, reset, depth, wrapper):
-        if cd_logic is None:
-            fifo = wrapper(SyncFIFOBuffered(width=8, depth=depth))
-
-            if reset is not None:
-                fifo = ResetInserter(reset)(fifo)
-        else:
-            assert isinstance(cd_logic, ClockDomain)
-
-            raw_fifo = DomainRenamer({
-                crossbar_side: "crossbar",
-                logic_side:    "logic",
-            })(AsyncFIFO(width=8, depth=depth))
-
-            fifo = wrapper(_AsyncFIFOWrapper(raw_fifo, cd_logic, reset))
-
-        return fifo
-
-    def get_out_fifo(self, n, depth=512, clock_domain=None, reset=None):
-        assert 0 <= n < 2
-        assert isinstance(self.out_fifos[n], _UnimplementedOUTFIFO)
-
-        fifo = self._make_fifo(crossbar_side="write",
-                               logic_side="read",
-                               cd_logic=clock_domain,
-                               reset=reset,
-                               depth=depth,
-                               wrapper=lambda fifo: _OUTFIFO(fifo,
-                                    skid_depth=3))
-        self.out_fifos[n] = fifo
-        return fifo
-
-    def get_in_fifo(self, n, depth=512, auto_flush=True, clock_domain=None, reset=None):
-        assert 0 <= n < 2
-        assert isinstance(self.in_fifos[n], _UnimplementedINFIFO)
-
-        fifo = self._make_fifo(crossbar_side="read",
-                               logic_side="write",
-                               cd_logic=clock_domain,
-                               reset=reset,
-                               depth=depth,
-                               wrapper=lambda fifo: _INFIFO(fifo,
-                                    asynchronous=clock_domain is not None,
-                                    auto_flush=auto_flush))
-        self.in_fifos[n] = fifo
-        return fifo
