@@ -5,6 +5,7 @@ import re
 from collections import deque
 
 from .aobject import *
+from .logging import dump_hex
 
 
 __all__ = ["ServerEndpoint", "ClientEndpoint"]
@@ -36,7 +37,8 @@ class ServerEndpoint(aobject, asyncio.Protocol):
             name, metavar=metavar, type=endpoint, nargs=nargs, default=default,
             help=help)
 
-    async def __init__(self, name, logger, sock_addr, queue_size=None):
+    async def __init__(self, name, logger, sock_addr, queue_size=None, *,
+                       deprecated_cancel_on_eof=False):
         assert isinstance(sock_addr, tuple)
 
         self.name    = name
@@ -70,6 +72,12 @@ class ServerEndpoint(aobject, asyncio.Protocol):
 
         self._read_paused = False
 
+        self._cancel_on_eof = deprecated_cancel_on_eof
+        if self._cancel_on_eof:
+            self._log(logging.WARNING,
+                "ServerEndpoint with cancel-on-EOF behavior is deprecated; please fix this applet "
+                "and submit a pull request (thanks in advance!)")
+
     def _log(self, level, message, *args):
         self._logger.log(level, self.name + ": " + message, *args)
 
@@ -88,6 +96,7 @@ class ServerEndpoint(aobject, asyncio.Protocol):
             self._log(logging.INFO, "closing old connection")
             self._transport.close()
             self._new_transport = transport
+        self.data_received(b"")
 
     def connection_lost(self, exc):
         peername = self._transport.get_extra_info("peername")
@@ -131,12 +140,18 @@ class ServerEndpoint(aobject, asyncio.Protocol):
     async def _refill(self):
         self._future = future = asyncio.Future()
         self._check_future()
-        self._buffer = await future
+        try:
+            self._buffer = await future
+        except BrokenPipeError:
+            self._buffer = None
         if self._buffer is None:
             self._buffer = b""
             self._log(logging.TRACE, "recv end-of-stream")
             self._recv_epoch += 1
-            raise asyncio.CancelledError
+            if self._cancel_on_eof:
+                raise asyncio.CancelledError
+            else:
+                raise EOFError
 
     async def recv(self, length=0):
         data = bytearray()
@@ -154,7 +169,7 @@ class ServerEndpoint(aobject, asyncio.Protocol):
             self._check_pushback()
             data += chunk
 
-        self._log(logging.TRACE, "recv <%s>", data.hex())
+        self._log(logging.TRACE, "recv <%s>", dump_hex(data))
         return data
 
     async def recv_until(self, separator):
@@ -180,7 +195,7 @@ class ServerEndpoint(aobject, asyncio.Protocol):
                 self._check_pushback()
                 self._buffer = None
 
-        self._log(logging.TRACE, "recv <%s%s>", data.hex(), separator.hex())
+        self._log(logging.TRACE, "recv <%s%s>", dump_hex(data), separator.hex())
         return data
 
     async def recv_wait(self):
@@ -190,8 +205,8 @@ class ServerEndpoint(aobject, asyncio.Protocol):
 
     async def send(self, data):
         data = bytes(data)
-        if self._send_epoch == self._recv_epoch:
-            self._log(logging.TRACE, "send <%s>", data.hex())
+        if self._transport is not None and self._send_epoch == self._recv_epoch:
+            self._log(logging.TRACE, "send <%s>", dump_hex(data))
             self._transport.write(data)
             return True
         else:
